+++ src/service.js
@@ -1,7 +1,7 @@
-import * as Common from "./handlers/common.js";
-import * as Features from "./handlers/features.js";
-import * as Tiles from "./handlers/tiles.js";
-import * as Styles from "./handlers/styles.js";
+import * as Common from "./routes/common.js";
+import * as Features from "./routes/features.js";
+import * as Tiles from "./routes/tiles.js";
+import * as Styles from "./routes/styles.js";
 
 class Service {
     constructor(fastify, baseurl) {
@@ -20,4 +20,3 @@ Object.assign(Service.prototype, {
 });
 
 export { Service }
-
